aws-lambda/enhance-images/requirements.txt
# AWS Lambda dependencies for image enhancement
boto3>=1.34.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0

